refactor: rename files deprecated-service to retired-service
